# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a sample server for FOC project not Frog OOB OOB.",
        "title": "FOC",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api",
    "paths": {
...
        "/constants/get-district": {
            "get": {
                "description": "Get all Thai District.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Constants"
                ],
                "summary": "Get Thai District",
                "operationId": "get-district",
                "responses": {
                    "200": {
                        "description": "District",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.District"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "District not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/constants/get-province": {
            "get": {
                "description": "Get all Thai Province.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Constants"
                ],
                "summary": "Get Thai Province",
                "operationId": "get-province",
                "responses": {
                    "200": {
                        "description": "Province",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.Province"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Province not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/constants/get-sku": {
            "post": {
                "description": "get sku by warehouse, location and site.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Constants"
                ],
                "summary": "get sku by warehouse, location and site",
                "operationId": "get-sku",
                "parameters": [
                    {
                        "description": "Get Sku",
                        "name": "get-sku",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/entity.WarehouseLocation"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Get sku by warehouse, location and site",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.GetSKU"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "sku not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/constants/get-sub-district": {
            "get": {
                "description": "Get all Thai SubDistrict.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Constants"
                ],
                "summary": "Get Thai SubDistrict",
                "operationId": "get-sub-district",
                "responses": {
                    "200": {
                        "description": "SubDistrict",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.SubDistrict"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "SubDistrict not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/constants/get-type-foc": {
            "get": {
                "description": "Get all Type foc.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Constants"
                ],
                "summary": "Get Type foc",
                "operationId": "get-type-foc",
                "responses": {
                    "200": {
                        "description": "Type FOC",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.TypeFoc"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Type Foc not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/constants/get-type-logistic": {
            "get": {
                "description": "Get all Type logistic.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Constants"
                ],
                "summary": "Get Type logistic",
                "operationId": "get-type-logistic",
                "responses": {
                    "200": {
                        "description": "Type Logistic",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.TypeLogistic"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Type Foc not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/constants/get-type-req": {
            "get": {
                "description": "Get all Type req.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Constants"
                ],
                "summary": "Get Type req",
                "operationId": "get-type-req",
                "parameters": [
                    {
                        "type": "string",
                        "description": "input Permission ID",
                        "name": "permissionID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Type Req",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/entity.TypeReqPermission"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "Type Req not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
        "/constants/get-warehouse-location": {
            "get": {
                "description": "get warehouse \u0026 location by user's id.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Constants"
                ],
                "summary": "get warehouse \u0026 location by user's id",
                "operationId": "get-warehouse-location",
                "parameters": [
                    {
                        "type": "string",
                        "description": "User's ID",
                        "name": "userID",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Warehouse Location by user'ID",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/api.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "result": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/service.SiteWarehouseLocations"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "404": {
                        "description": "warehoues location not found",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/api.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "api.Response": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string"
                },
                "result": {},
                "success": {
                    "type": "boolean"
                }
            }
        },
...
        "entity.District": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "nameEN": {
                    "type": "string"
                },
                "nameTH": {
                    "type": "string"
                },
                "provinceCode": {
                    "type": "integer"
                }
            }
        },
        "entity.GetSKU": {
            "type": "object",
            "properties": {
                "qtyAvailable": {
                    "type": "integer"
                },
                "sizeID": {
                    "type": "integer"
                },
                "sizeName": {
                    "type": "string"
                },
                "sku": {
                    "type": "string"
                },
                "skuName": {
                    "type": "string"
                }
            }
        },
        "entity.Province": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "nameEN": {
                    "type": "string"
                },
                "nameTH": {
                    "type": "string"
                }
            }
        },
        "entity.SubDistrict": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "districtCode": {
                    "type": "integer"
                },
                "nameEN": {
                    "type": "string"
                },
                "nameTH": {
                    "type": "string"
                },
                "zipCode": {
                    "type": "string"
                }
            }
        },
        "entity.TypeFoc": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "journalType": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "site": {
                    "type": "string"
                },
                "statusEN": {
                    "type": "string"
                },
                "statusTH": {
                    "type": "string"
                },
                "typeValue": {
                    "type": "integer"
                },
                "warehouse": {
                    "type": "string"
                }
            }
        },
        "entity.TypeLogistic": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "logisticType": {
                    "type": "integer"
                },
                "nameEN": {
                    "type": "string"
                },
                "nameTH": {
                    "type": "string"
                }
            }
        },
        "entity.TypeReqPermission": {
            "type": "object",
            "properties": {
                "journalType": {
                    "type": "string"
                },
                "location": {
                    "type": "string"
                },
                "nameTH": {
                    "type": "string"
                },
                "permissionID": {
                    "type": "string"
                },
                "reqType": {
                    "type": "string"
                },
                "site": {
                    "type": "string"
                },
                "typeValue": {
                    "type": "integer"
                },
                "warehouse": {
                    "type": "string"
                }
            }
        },
        "entity.WarehouseLocation": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "string"
                },
                "site": {
                    "type": "string"
                },
                "warehouse": {
                    "type": "string"
                }
            }
        },
...
        "service.SiteWarehouseLocations": {
            "type": "object",
            "properties": {
                "site": {
                    "type": "string"
                },
                "warehouseLocs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/service.WarehouseSection"
                    }
                }
            }
        },
        "service.WarehouseSection": {
            "type": "object",
            "properties": {
                "location": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "warehouse": {
                    "type": "string"
                }
            }
        }
    }
}